WIP IPBanningModule
